requirements.txt
pyautogui>=0.9.54
pytesseract>=0.3.10
Pillow>=10.0.0
PyYAML>=6.0
mss>=9.0.1
python-dotenv>=1.0.0
